% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvp.R
\name{pvp}
\alias{pvp}
\title{Plot and compare asreml model variance parameters}
\usage{
pvp(curr, prev = NULL, logt = FALSE, size = 4)
}
\arguments{
\item{curr}{object of class \code{asreml}: the current
\pkg{asreml} model, whose variance parameters are to be plotted.}

\item{prev}{(optional) object of class \code{asreml}: the previous
\pkg{asreml} model, whose variance parameters are to be comapred
to the current \pkg{asreml} model \code{curr}.}

\item{logt}{logical: log-transform the gamma axis? (Default
\code{FALSE}).}

\item{size}{numeric: point size parameter (default 4) see
\code{\link{geom_point}} in package \pkg{ggplot2}}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot variance parameters for one or two \pkg{asreml} models.
}
\author{
Alexander Zwart (alec.zwart at csiro.au) and Alex Whan
(alex.whan at csiro.au)
}

