% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeplot.R
\name{repeplot}
\alias{repeplot}
\title{Plot \pkg{asreml} random model parameters}
\usage{
repeplot(curr, prev = NULL, gammas = TRUE, size = 6, lwd = 2,
  zline = TRUE, rlines = TRUE, bw = FALSE, pointcol = ifelse(bw,
  "grey50", "#0072B2"), altcol = ifelse(bw, "black", "#E69F00"))
}
\arguments{
\item{curr}{\code{\link{asreml}} object: The 'current'
\code{\link{asreml}} model for which the estimated random effects
parameters are to be plotted.}

\item{prev}{\code{\link{asreml}} object, or \code{NULL}: If a
'previous' \code{\link{asreml}} model is specified via
\code{prev}, the plot will compare random effects parameters
between the previous and current models.  Default is \code{NULL}.
See the Details section for more.}

\item{gammas}{logical: Should the plot display 'gamma' values or
actual components?  Default is \code{TRUE}. See the Details
section for more.}

\item{size}{numeric: The size parameter for plotting points.
Equivalent to the \code{\link{geom_point}} '\code{size}'
parameter.}

\item{lwd}{numeric: The width parameter for line segments
indicating changes in parameters common to the current and
previous models.  Equivalent to the \code{\link{geom_segment}}
'\code{size}' parameter.}

\item{zline}{logical: Display a grey vertical line (the 'zero
line') passing through the x-axis origin? Default is \code{TRUE}.}

\item{rlines}{logical: Display dashed vertical line(s) passing
through the residual variance parameter estimate(s)? Default is
\code{TRUE}.}

\item{bw}{logical: Produce a black-and-white version of the plot?
Default is FALSE.  If \code{bw=TRUE}, parameters present in both
models are plotted as grey points and parameters in only one model
are plotted as unfilled circles.}

\item{pointcol}{Specifies the point colour for parameters present
in both models.  Default is \code{"#0072B2"} (a shade of blue), or
\code{"grey50"} when \code{bw=TRUE}.}

\item{altcol}{Specifies the point colour for parameters present in
only one model.  Default is \code{"#E69F00"} (a shade of orange)
when \code{bw=FALSE}.  When \code{bw=TRUE}, this parameter
controls the colour of the point \emph{outline}, and the default
value is \code{"black"}.}
}
\value{
A \pkg{ggplot2}-package \code{ggplot} object.
}
\description{
Function \code{repeplot} plots the random effects parameter
estimates from one or two fitted \code{\link{asreml}} models.  If
two models (termed the 'current' and 'previous' models) are
specified, the plot compares the parameter estimates from the
current model to those from the previous.
}
\details{
Function \code{repeplot} plots the random effects parameter
estimates from one or two fitted \code{\link{asreml}} model
objects.  The random effects parameter labels that are used are
those given in the \code{varcomp} component of the object returned
by \code{\link{summary.asreml}}.  'Gamma' values (see below), or
unscaled parameter values can be plotted.  Residual variance
parameters are highlighted on the plot via vertical dashed lines
when \code{rlines} is \code{TRUE} (the default), and the position
of the origin is highlighted by a vertical grey line when
\code{zline} is \code{TRUE} (the default).

Regarding 'gamma' values: when a linear mixed model contains a
single residual variance \code{\link{asreml}} will parameterise
other variances in the model \emph{relative} to the residual
variance.  The resulting scale factors are referred to as 'gamma'
parameters.  The \code{varcomp} component of the object returned
by \code{\link{summary.asreml}} reports these gamma values, along
with the unscaled component values, and \code{repeplot} allows you
to plot either, via the argument \code{gammas} (default value
\code{TRUE}).  Note that no such scaling is performed when a model
contains more than one residual variance parameter, and the
reported gamma estimates are identical to the unscaled component
estimates.  For further details on how \code{\link{asreml}}
parameterises the linear mixed model, see the \pkg{asreml} package
manual.

Parameters that lie 'on the boundary' (and hence indicated by
"Boundary" in the \code{varcomp$condition} component of the
summary object), are highlighted on the plot by a short vertical
bar through the plotted parameter.

When two models are specified in the \code{repeplot} call, these are
termed the 'current' and 'previous' models. \code{repeplot} aims to
highlight the changes in the parameter estimates in moving from
the previous to the current model.

As such, parameters that are present in one model and not the
other are plotted in a different colour to those parameters
present in both models.  Parameters present in the \emph{previous}
model, but not the \emph{current} model, are also overplotted with
a cross, highlighting their absence from the current model.

For parameters present in both models, the values for the current
model are plotted, and horizontal line segments are used to
indicate how these parameters have changed between the previous
and current models.  Note that these line segments are plotted
\emph{behind} the corresponding points, so if parameter changes
are very small, the line segments may be too short to be seen.

Caveat: note that a change in the random effects model
specification between 'previous' and 'current' models, can
sometimes result in an equivalent parameter being given
\emph{different labels} by \code{\link{asreml}} between the two
models.  \code{repeplot} cannot detect such cases, and hence will
plot the affected parameter as two distinct parameters - one
removed from the previous model, the other added to the current.

Not all random effects parameters are variances, and as such may
have different scales to the variances.  Autocorrelation
estimates, for example, lie between -1 and 1, and have no
relationship to the scale of the variances.  \code{repeplot} does
not attempt to distinguish between parameters on different scales,
and if you wish to 'zoom in' on a particular range of parameter
values to see the estimates clearly, you can do so using the usual
\pkg{ggplot2} \code{xlim} function syntax,
e.g. \code{repeplot(a1,a0) + xlim(-1,1)}.

\code{repeplot} requires that the \pkg{asreml}, \pkg{dplyr} and
\pkg{ggplot2} packages be installed.  \code{repeplot} returns a
\code{\link{ggplot}} object, and this object can be further
modified using the usual \pkg{ggplot2} package
syntax. \code{repeplot} uses the \code{\link{theme_bw}} theme by
default.
}
\author{
Alexander Zwart (alec.zwart at csiro.au)
}

