% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replot.R
\name{replot}
\alias{replot}
\title{Plot \pkg{asreml} random model parameters}
\usage{
replot(curr, prev = NULL, gammas = TRUE, size = 6, lwd = 2,
  tritanopia = FALSE, zline = TRUE, rlines = TRUE)
}
\arguments{
\item{curr}{asreml object: The 'current' asreml model for which
the estimated random effects parameters are to be plotted.}

\item{prev}{asreml object, or NULL.  If a 'previous' asreml model
is specified for \code{prev}, the plot will compare random effects
parameters between the previous and current models.  See the
details section for more.}

\item{gammas}{logical: Should the plot display 'gamma' values or}

\item{size}{numeric: The size parameter for plotting points.
Equivalent to the \code{\link{geom_point}} '\code{size}'
parameter.}

\item{lwd}{numeric: width parameter for the line segments that
indicate changes in parameters common to the current and previous
models.  Equivalent to the \code{\link{geom_segment}}
'\code{size}' parameter.}

\item{tritanopia}{logical: The default colors used by \code{replot}
should (I hope!) suit non-colourblind users, as well as those with
from protanopia and deuteranopia. Tritanopia-friendly colors can
be obtained by setting \code{tritanopia=TRUE}.  Default is FALSE.}

\item{zline}{logical: Display a grey vertical line (the 'zero
line') passing through the x-axis origin? Default is TRUE.}

\item{rlines}{logical: Display dashed vertical lines passing
through the residual variance parameter estimates? Default is
TRUE.}
}
\value{
A \pkg{ggplot2}-package \code{ggplot} object.
}
\description{
Plot the random effects parameter estimates from one or two fitted
\code{\link{asreml}} models.  If two models (termed the 'current' and
'previous' models) are specified, the plot compares the parameter
estimates from the current model to those from the previous.
}
\details{
The \code{asreml} method \code{replot} plots the random effects
parameter estimates from one or two fitted \code{link{asreml}}
model objects.  The random effects parameter labels that are used
are those given in the \code{varcomp} component of the object
returned by \code{\link{summary.asreml}}.  Gamma values (the
default), or unscaled parameter values can be plotted.  Residual
variance parameters are highlighted on the plot via vertical
dashed lines, and the position of the origin is highlighted by a
vertical grey line.

Parameters that lie 'on the boundary' (and hence indicated by
"Boundary" in the \code{varcomp$condition} component of the asreml
summary object), are highlighted on the plot by a short vertical
bar through the plotted parameter.

When two models are specified in the \code{replot} call, these are
termed the 'current' and 'previous' models, and \code{replot} aims
to show the changes in the parameter estimates in going from the
previous to the current model.

Hence, parameters that are present in one model and not the other
are plotted in a different colour to those parameters present in
both models.  Parameters present in the \emph{previous} model, but
not the \emph{current} model, are also overplotted with a cross,
highlighting their absence from the current model.

For parameters present in both models, the values for the current
model are plotted, and horizontal line segments are used to
indicate how these parameters have changed between the previous and
current models.  Note that if parameter changes are very small,
the line segments may be too short to be seen.

Caveat: note that a change in the random effects model
specification between 'previous' and 'current' models, can
sometimes result in an equivalent parameter being given different
labels by \code{\link{asreml}} between the models.  \code{replot}
cannot detect such cases, and hence will plot the affected
parameter as two distinct parameters - one removed from the
previous model, the other added to the current.

\code{replot} requires that the \pkg{dplyr} and \pkg{ggplot2}
packages be installed.  \code{replot} returns a \code{ggplot}
object, and this object can be further modified using the usual
\pkg{ggplot2} package syntax.  For example (assuming asreml model
objects '\code{a1}' and '\code{a0}'), one can change the x-axis
scale of the object via, say, '\code{replot(a1,a0) + xlim(-1,1)}'.
}
\author{
Alexander Zwart (alec.zwart at csiro.au)
}

